refactor(packages/eg-walker): split columnar-codec into per-column modules
Extract the three independent column groups (operation runs, parent
overrides, id runs) from the 581-line flat file into dedicated modules
under graph/columnar-codec/. Each column's encode/write/read/decode
functions are standalone pure functions; ColumnarEventGraphCodec in
index.ts becomes a thin orchestrator. No import paths changed — TS
resolves the folder via index.ts automatically.
